restructure all react env webpack configs
