+++ scopes/react/react/webpack/webpack.config.component.dev.ts
@@ -1,0 +1,77 @@
+import path from 'path';
+import { Configuration } from 'webpack';
+import { ComponentID } from '@teambit/component-id';
+
+// This is the production and development configuration.
+// It is focused on developer experience, fast rebuilds, and a minimal bundle.
+// eslint-disable-next-line complexity
+export default function (fileMapPath: string, workDir: string): Configuration {
+  return {
+    module: {
+      rules: [
+        {
+          test: /\.js$/,
+          enforce: 'pre',
+          // limit loader to files in the current project,
+          // to skip any files linked from other projects (like Bit itself)
+          include: path.join(workDir, 'node_modules'),
+          // only apply to packages with componentId in their package.json (ie. bit components)
+          descriptionData: { componentId: (value) => !!value },
+          use: [require.resolve('source-map-loader')],
+        },
+        {
+          test: /\.js$/,
+          // limit loader to files in the current project,
+          // to skip any files linked from other projects (like Bit itself)
+          include: path.join(workDir, 'node_modules'),
+          // only apply to packages with componentId in their package.json (ie. bit components)
+          descriptionData: { componentId: ComponentID.isValidObject },
+          use: [
+            {
+              loader: require.resolve('babel-loader'),
+              options: {
+                babelrc: false,
+                configFile: false,
+                plugins: [
+                  // for component highlighting in preview.
+                  [require.resolve('@teambit/react.babel.bit-react-transformer')],
+                ],
+                // turn off all optimizations (only slow down for node_modules)
+                compact: false,
+                minified: false,
+              },
+            },
+          ],
+        },
+        {
+          test: /\.(mjs|js|jsx|tsx|ts)$/,
+          // TODO: use a more specific exclude for our selfs
+          exclude: [/node_modules/, /dist/],
+          include: workDir,
+          resolve: {
+            fullySpecified: false,
+          },
+          loader: require.resolve('babel-loader'),
+          options: {
+            babelrc: false,
+            configFile: false,
+            presets: [
+              // Preset includes JSX, TypeScript, and some ESnext features
+              require.resolve('babel-preset-react-app'),
+            ],
+            plugins: [
+              // require.resolve('react-refresh/babel'),
+              // for component highlighting in preview.
+              [
+                require.resolve('@teambit/react.babel.bit-react-transformer'),
+                {
+                  componentFilesPath: fileMapPath,
+                },
+              ],
+            ],
+          },
+        },
+      ],
+    },
+  };
+}
